+++ src/extension.ts
@@ -1,4 +1,4 @@
-import { ExtensionContext, StatusBarAlignment, StatusBarItem, window } from "vscode";
+import { ExtensionContext, Position, StatusBarAlignment, StatusBarItem, window } from "vscode";
 
 const StatusBarId = "cursor_pos";
 const Alignment = StatusBarAlignment.Right;
@@ -11,6 +11,7 @@ export function activate(context: ExtensionContext) {
 
     context.subscriptions.push(statusBarItem);
     context.subscriptions.push(window.onDidChangeTextEditorSelection(updateStatusBarItem));
+    context.subscriptions.push(window.onDidChangeActiveTextEditor(updateStatusBarItem));
 
     // Update once on init.
     updateStatusBarItem();
@@ -18,11 +19,24 @@ export function activate(context: ExtensionContext) {
 
 function updateStatusBarItem() {
     const activeEditor = window.activeTextEditor;
-    if (activeEditor === undefined) {
+    // Don't show the status bar item if there's no editor
+    // or if more than one selection is active.
+    if (activeEditor === undefined || activeEditor.selections.length !== 1) {
         statusBarItem.hide();
     } else {
-        const offset = activeEditor.document.offsetAt(activeEditor.selection.active);
-        statusBarItem.text = `Offset ${offset}`;
+        const selection = activeEditor.selection;
+        const start = selection.anchor;
+        const end = selection.active;
+
+        const startOffset = activeEditor.document.offsetAt(start);
+        const endOffset = activeEditor.document.offsetAt(end);
+
+        if (startOffset !== endOffset) {
+            statusBarItem.text = `Offset ${startOffset}-${endOffset}`;
+        } else {
+            statusBarItem.text = `Offset ${startOffset}`;
+        }
+
         statusBarItem.show();
     }
 }