+++ CHANGELOG.md
@@ -8,4 +8,6 @@ Initial release
 
 ### Added
 
-- Status bar item that displays the current absolute, 0-indexed position of the cursor within the current document.
+- Status bar item that displays the absolute, 0-indexed position of the cursor within the current document.
+  - Displays the start and end offsets of a selection.
+  - Does not appear when more than one cursor or selection is active.